backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pymysql==1.1.1
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
celery==5.4.0
redis==5.1.1
firebase-admin==6.5.0
boto3==1.35.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
